First Attempt

+++ payCalculator.py
@@ -4,9 +4,8 @@ def calculatePay():
     # This first line is provided for you
     hrs = input("Enter Hours:")
     rate = input("Enter Rate:")
-    pay = float(hrs)*float(rate)
+    pay = int(hrs)*float(rate)
     print(pay)
-    
     
     
     # end assignment
